top(1): style and related
- style(9)
- remove now-defunct comments
- remove getuid check for low delay
- expand range of format_k


git-svn-id: svn+ssh://svn.freebsd.org/base/head@334989 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.bin/top/utils.h
@@ -19,7 +19,7 @@ int digits(int);
 const char * const *argparse(char *, int *);
 long percentages(int, int *, long *, long *, long *);
 char *format_time(long);
-char *format_k(int);
+char *format_k(long);
 char *format_k2(unsigned long long);
 int string_index(const char *string, const char * const *array);
 int find_pid(pid_t pid);
